Split test_template_message.py into two files
